project/requirements.txt
django==1.8.6
simplejson
gunicorn
psycopg2
Pillow==2.9.0
django-imagekit
django-flat-theme

djangorestframework
markdown
django-filter
django-cors-headers
django-rest-auth
djangorestframework-gis

Celery
django-celery

apscheduler
requests